+++ src/protocols/disconnect.rs
@@ -50,7 +50,7 @@ where
         self.node().tasks.lock().push(disconnect_task);
 
         // register the Disconnect handler with the Node
-        let hdl = ProtocolHandler(from_node_sender);
+        let hdl = Box::new(ProtocolHandler(from_node_sender));
         assert!(
             self.node().protocols.disconnect_handler.set(hdl).is_ok(),
             "the Disconnect protocol was enabled more than once!"